Create decorator and use it for reverse_segment #64
This change includes a test_editor refactoring. To split api and integration. So, api tests can be also split into views.
